the core now builds without warnings

+++ core/JMXRect.mm
@@ -176,10 +176,10 @@ v8::Handle<v8::Value> JMXRectJSConstructor(const v8::Arguments& args)
     HandleScope handleScope;
     //v8::Locker locker;
     v8::Persistent<FunctionTemplate> objectTemplate = [JMXRect jsObjectTemplate];
-    int x = 0;
-    int y = 0;
-    int w = 0;
-    int h = 0;
+    int64_t x = 0;
+    int64_t y = 0;
+    int64_t w = 0;
+    int64_t h = 0;
     if (args.Length() >= 2) {
         x = args[0]->IntegerValue();
         y = args[1]->IntegerValue();
